Safari-609.1.11.1/LayoutTests/js/dom/array-float-delete-expected.txt

This test checks for regression against 6234: Can delete array index property incorrectly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a[1] is "before"
PASS successfullyParsed is true

TEST COMPLETE
